Add a couple rooms, changed gen code
We now select a random room to place, and a random door on that room to connect to. It then does some rotation based of the rotation of the doors in the base room and new room, then offsets the position of the starting door by the new door to get the room's position. Finally, it deletes the room.